backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.42.0
motor==3.7.1
pydantic==2.7.4
pydantic-settings==2.2.1
python-jose==3.3.0
bcrypt==4.1.2
redis==5.0.1
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.1.0
httpx==0.28.1
PyMuPDF==1.23.1
numpy==1.26.4
faiss-cpu==1.7.4
langchain==1.2.12
llama-index==0.9.1
sse-starlette==1.2.1
pydantic[email]==2.7.4
loguru==0.7.0
python-slugify==8.0.1
Pillow==10.0.0
pyotp==2.9.0
